Derived Annual station data

Derived data collected at fixed stations by year.

Derived_Annual_Stations

Format

A dataframe

station

The name of the station where data was collected

year

The year in which data was collected

temperature_in_air

Average air temperatures over annual scales at specific stations including Sable Island

denisty

Discrete density measurements over a range of depths

sea_temperature

Discrete sea temperature measurements over a range of depths

salinity

Discrete salinity measurements over a range of depths

integrated_sea_temperature_0_50

Averages of temperature measurements, integrated between 0 and 50 metres

integrated_salinity_0_50

Averages of salinity measurements, integrated between 0 and 50 metres

integrated_density_0_50

Averages of density measurements, integrated between 0 and 50 metres

temperature_0

Averages for temperature at 0 metres at Prince 5 (P5) and Halifax 2 (HL2) stations

temperature_90

Averages for temperature at 90 metres at Prince 5 (P5) and Halifax 2 (HL2) stations

sea_surface_temperature_from_moorings

Averages of sea surface temperatures recorded by mooring at Halifax, NS and St. Andrews, NB

sea_surface_height

Average of sea surface height at Yarmouth and Halifax NS

integrated_nitrate_0_50

Nitrate concentrations integrated between 0 and 50 metres

integrated_nitrate_50_150

Nitrate concentrations integrated between 50 and 150 metres

integrated_phosphate_0_50

Phosphate concentrations integrated between 0 and 50 metres

integrated_phosphate_50_150

Phosphate concentrations integrated between 50 and 150 metres

Details

The data can be cited as follows: Casault, B., Johnson, C., Devred, E., Head, E., Cogswell, A., and Spry, J. 2020. Optical, Chemical, and Biological Oceanographic Conditions on the Scotian Shelf and in the Eastern Gulf of Maine during 2018. DFO Can. Sci. Advis. Sec. Res. Doc. 2020/037. v + 66 p.

Hebert, D., Pettipas, R., and Brickman, D. 2020. Physical Oceanographic Conditions on the Scotian Shelf and in the Gulf of Maine during 2018. DFO Can. Sci. Advis. Sec. Res. Doc. 2020/036 iv + 52 p.

Note

time_scale

Annual

regional_scale

Station

category

biological, phytoplankton, biochemical, physical