Derived Occupation Section data

Derived data collected along AZMP transects during individual occupations.

Derived_Occupations_Sections

Format

A dataframe

section

Section name where data was collected

station

The name of the station where data was collected

latitude

The latitude at which data was collected

longitude

The longitude at which data was collected

year

The year in which data was collected

month

The month in which data was collected (numeric)

day

The day on which data was collected

event_id

A unique identifier for the sampling event

cruise_id

A unique identifier for the mission/cruise

odf_filename

The original data file name

integrated_nitrate_0_50

Nitrate concentrations integrated between 0 and 50 metres

integrated_nitrate_50_150

Nitrate concentrations integrated between 50 and 150 metres

integrated_phosphate_0_50

Phosphate concentrations integrated between 0 and 50 metres

integrated_phosphate_50_150

Phosphate concentrations integrated between 50 and 150 metres

integrated_silicate_0_50

Silicate concentrations integrated between 0 and 50 metres

integrated_silicate_50_150

Silicate concentrations integrated between 50 and 150 metres

integrated_chlorophyll_0_100

Chlorophyll concentrations integrated between 0 and 100 metres

Details

The data can be cited as follows: Casault, B., Johnson, C., Devred, E., Head, E., Cogswell, A., and Spry, J. 2020. Optical, Chemical, and Biological Oceanographic Conditions on the Scotian Shelf and in the Eastern Gulf of Maine during 2018. DFO Can. Sci. Advis. Sec. Res. Doc. 2020/037. v + 66 p.

Note

time_scale

Occupation

regional_scale

Section

category

biological, biochemical