Zooplankton Annual Stations data

Zooplankton data collected at fixed stations by year.

Zooplankton_Annual_Stations

Format

A dataframe

station

The name of the station where data was collected

year

The year in which data was collected

calanus_finmarchicus_log10

Abundance of calanus finmarchicus with a log 10 transform

Acartia_log10

Abundance of acartia with a log 10 transform

Eurytemora_log10

Abundance of eurytemora with a log 10 transform

Temora_longicorns_log10

Abundance of temora longicornis with a log 10 transform

other_copepods_log10

Abundance of other copepods group with a log 10 transform

copepods_log10

Abundance of copepods group with a log 10 transform

non_copepods_log10

Abundance of non-copepods group with a log 10 transform

Arctic_Calanus_species_log10

Abundance of arctic calanus group with a log 10 transform

warm_offshore_copepods_log10

Abundance of warm offshore copepods group (Clausocalanus spp., Mecynocera clausi, and Pleuromamma borealis) with a log 10 transform

warm_shelf_copepods_log10

Abundance of warm shelf copepods group with a log 10 transform (the summer-fall copepods Paracalanus spp. and Centropages typicus)

calanus_hyperboreus_log10

Abundance of calanus hyperboreus with a log 10 transform

pseudocalanus_log10

Abundance of pseudocalanus with a log 10 transform

metridia_longa_log10

Abundance of metridia longa with a log 10 transform

metridia_lucens_log10

Abundance of metridia lucens with a log 10 transform

metridia_spp_log10

Abundance of metridia spp with a log 10 transform

microcalanus_spp_log10

Abundance of microcalanus with a log 10 transform

oithona_spp_log10

Abundance of oithona species with a log 10 transform

oithona_similis_log10

Abundance of oithona similis with a log 10 transform

oithona_atlantica_log10

Abundance of oithona atlantica with a log 10 transform

paracalanus_spp_log10

Abundance of paracalanus with a log 10 transform

centropages_typicus_log10

Abundance of centropages typicus with a log 10 transform

centropages_spp_log10

Abundance of centropages species with a log 10 transform

zooplankton_meso_dry_weight

Dry weight of meso zooplankton

zooplankton_total_wet_weight

Wet weight of all zooplankton from sampling

Details

The data can be cited as follows: Casault, B., Johnson, C., Devred, E., Head, E., Cogswell, A., and Spry, J. 2020. Optical, Chemical, and Biological Oceanographic Conditions on the Scotian Shelf and in the Eastern Gulf of Maine during 2018. DFO Can. Sci. Advis. Sec. Res. Doc. 2020/037. v + 66 p.

Note

time_scale

Annual

regional_scale

Station

category

biological, zooplankton