Zooplankton Occupation Section data

Zooplankton data collected along AZMP transects during individual occupations.

Zooplankton_Occupations_Sections

Format

A dataframe

section

Section name where data was collected

station

The name of the station where data was collected

latitude

The latitude at which data was collected

longitude

The longitude at which data was collected

year

The year in which data was collected

season

The season in which data was collected (Northern Hemisphere)

month

The month in which data was collected (numeric)

day

The day on which data was collected

event_id

A unique identifier for the sampling event

cruise_id

A unique identifier for the mission/cruise

sample_id

A unique identifier for the sample

depth

The depth at which data was actually collected

calanus_finmarchicus_abundance

Abundance of calanus finmarchicus

calanus_hyperboreus_abundance

Abundance of calanus hyperboreus

calanus_glacialis_abundance

Abundance of calanus glacialis

pseudocalanus_abundance

Abundance of pseudocalanus

metridia_longa_abundance

Abundance of metridia longa

metridia_lucens_abundance

Abundance of metridia lucens

metridia_spp_abundance

Abundance of metridia species

temora_spp_abundance

Abundance of temora species

microcalanus_spp_abundance

Abundance of microcalanus

oithona_spp_abundance

Abundance of oithona species

oithona_similis_abundance

Abundance of oithona similis

oithona_atlantica_abundance

Abundance of oithona atlantica

paracalanus_spp_abundance

Abundance of paracalanus

centropages_typicus_abundance

Abundance of centropages typicus

centropages_spp_abundance

Abundance of centropages species

scolecithricella_minor_abundance

Abundance of scolecithricella minor

larvacae_abundance

Abundance of larvacae group

gastropoda_abundance

Abundance of gastropoda group

bivalvia_abundance

Abundance of bivalvia group

euphasiacea_abundance

Abundance of euphasiacea group

zooplankton_abundance

Abundance of zooplankton at fixed stations

zooplankton_meso_dry_weight

Dry weight of meso zooplankton

zooplankton_meso_wet_weight

Wet weight of meso zooplankton

zooplankton_macro_wet_weight

Wet weight of macro zooplankton

zooplankton_total_wet_weight

Wet weight of all zooplankton from sampling

calanus_abundance_stage_c1

Stage seperated calanus abundance (stage c1)

calanus_abundance_stage_c2

Stage seperated calanus abundance (stage c2)

calanus_abundance_stage_c3

Stage seperated calanus abundance (stage c3)

calanus_abundance_stage_c4

Stage seperated calanus abundance (stage c4)

calanus_abundance_stage_c5

Stage seperated calanus abundance (stage c5)

calanus_abundance_stage_c6

Stage seperated calanus abundance (stage c6)

Details

The data can be cited as follows: Casault, B., Johnson, C., Devred, E., Head, E., Cogswell, A., and Spry, J. 2020. Optical, Chemical, and Biological Oceanographic Conditions on the Scotian Shelf and in the Eastern Gulf of Maine during 2018. DFO Can. Sci. Advis. Sec. Res. Doc. 2020/037. v + 66 p.

Note

time_scale

Occupation

regional_scale

Section

category

biological, zooplankton