Zooplankton seasonal broadscale data

A dataframe containing seasonal broadscale zooplankton data.

Zooplankton_Seasonal_Broadscale

Format

A dataframe containing 40 observations of 6 variables:

region

The region in which data was collected

year

The year in which data was collected

season

The season in which data was collected (Northern Hemisphere)

Calanus_finmarchicus_log10

Abundance of Calanus finmarchicus with a log10 transform

zooplankton_meso_dry_weight

Dry weight of meso zooplankton

zooplankton_total_wet_weight

Wet weight of all zooplankton sampled

@note

time_scale

Seasonal

regional_scale

Broadscale

category

Zooplankton, biological, biochemical

Details

The data can be cited as follows: Casault, B., Johnson, C., Devred, E., Head, E., Cogswell, A., and Spry, J. 2020. Optical, Chemical, and Biological Oceanographic Conditions on the Scotian Shelf and in the Eastern Gulf of Maine during 2018. DFO Can. Sci. Advis. Sec. Res. Doc. 2020/037. v + 66 p.