plot_availability

This generates a score card of samples per month based on areaType, areaName, and parameter.

plot_availability(
  areaType = NULL,
  areaName = NULL,
  parameters = NULL,
  fuzzyParameters = FALSE
)

Arguments

areaType

default is NULL. Indicates which type of area is of interest. The options are section, station, or area.

areaName

default is NULL. Indicates which area is of interest. The options are dependent on the areaType, however, a error message will let the user know which options are available.

parameters

default is NULL. Indicates which parameter is of interest. This is dependent on areaType and areaName, however, an error message will let the user know which options are available.

fuzzyParameters

default is TRUE. By default, any discovered parameters that match values within parameters will be returned. For example, parameter="nitrate" will return fields such as "integrated_nitrate_0_50", "integrated_nitrate_50_150", and "nitrate". If you want exact matches only, set fuzzyParameters = F.

Author

Jaimie Harbin jaimie.harbin@dfo-mpo.gc.ca

Mike McMahon, Mike.McMahon@dfo-mpo.gc.ca

Examples

if (FALSE) { # \dontrun{
plot_availability(areaType="station", areaName="P5", parameter="nitrate")
} # }